requirements.txt
anyio==3.6.2
azure-core==1.26.1
azure-devops==6.0.0b4      
azure-storage-blob==12.14.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1  
colorama==0.4.6
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
fastapi==0.88.0
fonttools==4.38.0
idna==3.4
ipykernel==6.16.2
ipython==7.34.0
isodate==0.6.1
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==4.12.0
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
msrest==0.6.21
nest-asyncio==1.5.6
numpy==1.21.6
oauthlib==3.2.2
packaging==22.0
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
psutil==5.9.4
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
starlette==0.22.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5